Create ROXY
Stick to healthy home-made meals. Yea, this looks like a big task for most, but think of it this way; you are surer of what you prepare yourself! Why not try to cook a “DIY” (do it yourself) healthy meal. Fill your plate or lunch box with a portion of veggies, fruits, and protein. Eat breakfast and keep yourself replenished by drinking water always.